// src/app/shared/models/import/ImportSummaryData.ts
export interface ImportSummaryData {
  importPadOs: string;
  ImportOsLiab: string;
  impAccChgAmount: string;
  impPendingLc: string;
  impPayChgAmount: string;
  impCommAmount: string;
  importPayment: string;
  impSwftChgAmount: string;
  importAmount: string;
  importLcOpen: string;
  importVatAmount: string;
  importTaxAmount: string;
  importPadAmt:string;
  importBillAmt:string;
  importOpChg:string;
}